Implemented CSI-RS transmission and reception.

See merge request aburger/openairinterface5g!270

+++ openair1/PHY/defs_nr_sl_UE.h
@@ -132,22 +132,22 @@ typedef struct SL_NR_UE_PSSCH {
   int16_t rsrp_dBm_per_RE;
 
   // STATS - CRC Errors observed during PSSCH reception (per HARQ round)
-  uint16_t rx_errors[8];
+  uint32_t rx_errors[8];
 
   // STATS - CRC Errors observed during PSSCH SCI2 reception
-  uint16_t rx_sci2_errors;
+  uint32_t rx_sci2_errors;
 
   // STATS - Receptions with CRC OK
-  uint16_t rx_ok;
+  uint32_t rx_ok;
 
   // STATS - Receptions with CRC OK
-  uint16_t rx_sci2_ok;
+  uint32_t rx_sci2_ok;
 
   // STATS - transmissions of PSSCH by the UE
-  uint16_t num_pssch_tx;
+  uint32_t num_pssch_tx;
 
   // STATS - transmissions of PSSCH by the UE
-  uint16_t num_pssch_sci2_tx;
+  uint32_t num_pssch_sci2_tx;
 } SL_NR_UE_PSSCH_t;
 
 typedef struct SL_NR_UE_PSCCH {
@@ -158,10 +158,10 @@ typedef struct SL_NR_UE_PSCCH {
   int16_t rsrp_dBm_per_RE;
 
   // STATS - Receptions with CRC OK
-  uint16_t rx_ok;
+  uint32_t rx_ok;
 
   // STATS - transmissions of PSBCH by the UE
-  uint16_t num_pscch_tx;
+  uint32_t num_pscch_tx;
 
 } SL_NR_UE_PSCCH_t;
 
@@ -173,19 +173,19 @@ typedef struct SL_NR_UE_PSBCH {
   int16_t rsrp_dBm_per_RE;
 
   // STATS - CRC Errors observed during PSBCH reception
-  uint16_t rx_errors;
+  uint32_t rx_errors;
 
   // STATS - Receptions with CRC OK
-  uint16_t rx_ok;
+  uint32_t rx_ok;
 
   // STATS - transmissions of PSBCH by the UE
-  uint16_t num_psbch_tx;
+  uint32_t num_psbch_tx;
 
 } SL_NR_UE_PSBCH_t;
 
 typedef struct SL_NR_UE_PSFCH {
   // STATS - transmissions of PSFCH by the UE
-  uint16_t num_psfch_tx;
+  uint32_t num_psfch_tx;
 } SL_NR_UE_PSFCH_t;
 
 typedef struct sl_nr_ue_phy_params {
